+++ plugin.php
@@ -13,7 +13,7 @@
  * Plugin URI:  https://wordpress.org/plugins/wp-plugin-dependencies
  * Description: Parses 'Requires Plugins' header, add plugin install dependencies tab, and information about dependencies.
  * Author: Andy Fragen, Colin Stewart, Paul Biron
- * Version: 1.10.0.5
+ * Version: 1.10.0.6
  * License: MIT
  * Network: true
  * Requires at least: 6.0
@@ -63,6 +63,7 @@ class Init {
 	 * @return void
 	 */
 	public function __construct() {
+		require_once __DIR__ . '/wp-admin/includes/plugin-install.php';
 		require_once __DIR__ . '/wp-admin/includes/class-wp-plugin-dependencies-2.php';
 
 		if ( ! WP_PLUGIN_DEPENDENCIES1_COMMITTED ) {